looker-actionhub-dev/requirements.txt
Flask==2.2.2
googleapis-common-protos==1.57.0
google-api-core
google-api-python-client
google-auth==2.14.1
google-cloud-bigquery==3.4.1
google-cloud-logging==3.5.0
grpc-google-iam-v1==0.12.4
grpcio==1.59.0
pandas==2.1.1
ptvsd==4.3.2 # Required for debugging.
pyarrow==10.0.1
pysftp==0.2.9
requests==2.28.1
waitress==2.1.2
boto3==1.26.147
Werkzeug==2.2.2
numpy==1.26.4
fsspec
s3fs
db-dtypes
google-ads==25.0.0